moved token registration tab from settings to charging stations menu
