+++ src/app/pages/charging-stations/registration-tokens/registration-token/registration-token.component.html
@@ -1,0 +1,67 @@
+<div class="main-content">
+    <div class="container-fluid">
+      <div class="row">
+        <div class="col-md-12">
+          <form [formGroup]="formGroup" class="form">
+            <div class="card card-profile card-testimonial">
+              <div class="card-header card-header-tabs card-header-primary">
+                <div class="nav-tabs-navigation">
+                  <div class="nav-tabs-wrapper">
+                    <ul class="nav nav-tabs" data-tabs="tabs">
+                      <li class="nav-item">
+                        <a class="nav-link active" data-toggle="tab" href="#profile">
+                          <mat-icon>link</mat-icon>
+                          <div class="ripple-container"></div>
+                          <span>{{description.value}}</span>
+                        </a>
+                      </li>
+                      <span class="toolbar-spacer"></span>
+                      <button mat-button (click)="save(formGroup.value)" [disabled]="!formGroup.valid">
+                        <mat-icon>save</mat-icon>
+                        <span>{{'general.save' | translate}}</span>
+                      </button>
+                      <button mat-icon-button (click)="close()">
+                        <mat-icon>close</mat-icon>
+                      </button>
+                    </ul>
+                  </div>
+                </div>
+              </div>
+              <div class="card-body">
+                <div class="tab-content">
+                  <div class="tab-pane active" id="profile">
+                    <div class="d-flex align-items-start flex-column mb-3">
+                      <mat-form-field class="col-md-12 form-group">
+                        <input [formControl]="description" appAutofocus matInput
+                          placeholder="{{'general.description' | translate}}" type="text">
+                        <mat-error *ngIf="description.errors?.required">{{"general.mandatory_field" | translate}}</mat-error>
+                        <mat-error *ngIf="description.errors?.maxlength">
+                          <div [translateParams]="{length: 100}" [translate]="'general.error_max_length'"></div>
+                        </mat-error>
+                      </mat-form-field>
+                      <mat-form-field class="col-md-12 col-lg-8 col-xl-6 form-group">
+                        <mat-datetimepicker #picker type="datetime" openOnFocus="false" mode="portrait" timeInterval="1">
+                        </mat-datetimepicker>
+                        <mat-datetimepicker-toggle [for]="picker" matSuffix></mat-datetimepicker-toggle>
+                        <input [formControl]="expirationDate" matInput placeholder="{{'general.expired_on' | translate}}"
+                          [matDatetimepicker]="picker" autocomplete="false">
+                      </mat-form-field>
+                      <mat-form-field class="col-md-12 form-group" *ngIf="isOrganizationComponentActive">
+                        <input matInput type="text" readonly=true placeholder="{{'chargers.site_area' | translate}}"
+                          class="form-field-popup" (click)="assignSiteArea()" [formControl]="siteArea"/>
+                        <button mat-button matSuffix mat-icon-button aria-label="Add" (click)="assignSiteArea()">
+                          <mat-icon>create</mat-icon>
+                        </button>
+                      </mat-form-field>
+                      <div class="clearfix"></div>
+                    </div>
+                  </div>
+                </div>
+              </div>
+            </div>
+          </form>
+        </div>
+      </div>
+    </div>
+  </div>
+
