+++ src/app/pages/charging-stations/registration-tokens/registration-tokens-table-data-source.ts
@@ -4,11 +4,11 @@ import { Router } from '@angular/router';
 import { TranslateService } from '@ngx-translate/core';
 import * as moment from 'moment';
 import { Observable } from 'rxjs';
-import { ComponentService } from 'services/component.service';
-import TenantComponents from 'types/TenantComponents';
 
+import { AuthorizationService } from '../../../services/authorization.service';
 import { CentralServerNotificationService } from '../../../services/central-server-notification.service';
 import { CentralServerService } from '../../../services/central-server.service';
+import { ComponentService } from '../../../services/component.service';
 import { DialogService } from '../../../services/dialog.service';
 import { MessageService } from '../../../services/message.service';
 import { SpinnerService } from '../../../services/spinner.service';
@@ -27,6 +27,7 @@ import { ButtonAction, RestResponse } from '../../../types/GlobalType';
 import { RegistrationToken } from '../../../types/RegistrationToken';
 import { SiteArea } from '../../../types/SiteArea';
 import { ButtonType, TableActionDef, TableColumnDef, TableDef, TableFilterDef } from '../../../types/Table';
+import TenantComponents from '../../../types/TenantComponents';
 import { User } from '../../../types/User';
 import { Utils } from '../../../utils/Utils';
 import { RegistrationTokenStatusComponent } from './registration-token/registration-token-status.component';
@@ -37,12 +38,12 @@ export class RegistrationTokensTableDataSource extends TableDataSource<Registrat
   private readonly isOrganizationComponentActive: boolean;
   private deleteAction = new TableDeleteAction().getActionDef();
   private revokeAction = new TableRevokeAction().getActionDef();
-  private copySOAP15Action = new TableCopyAction('settings.charging_station.ocpp_15_soap').getActionDef();
-  private copySOAP16Action = new TableCopyAction('settings.charging_station.ocpp_16_soap').getActionDef();
-  private copyJSON16Action = new TableCopyAction('settings.charging_station.ocpp_16_json').getActionDef();
-  private copySOAP15SecureAction = new TableCopyAction('settings.charging_station.ocpp_15_soap_secure').getActionDef();
-  private copySOAP16SecureAction = new TableCopyAction('settings.charging_station.ocpp_16_soap_secure').getActionDef();
-  private copyJSON16SecureAction = new TableCopyAction('settings.charging_station.ocpp_16_json_secure').getActionDef();
+  private copySOAP15Action = new TableCopyAction('chargers.connections.ocpp_15_soap').getActionDef();
+  private copySOAP16Action = new TableCopyAction('chargers.connections.ocpp_16_soap').getActionDef();
+  private copyJSON16Action = new TableCopyAction('chargers.connections.ocpp_16_json').getActionDef();
+  private copySOAP15SecureAction = new TableCopyAction('chargers.connections.ocpp_15_soap_secure').getActionDef();
+  private copySOAP16SecureAction = new TableCopyAction('chargers.connections.ocpp_16_soap_secure').getActionDef();
+  private copyJSON16SecureAction = new TableCopyAction('chargers.connections.ocpp_16_json_secure').getActionDef();
   private copyUrlAction: TableActionDef;
 
   constructor(
@@ -55,6 +56,7 @@ export class RegistrationTokensTableDataSource extends TableDataSource<Registrat
     private componentService: ComponentService,
     private centralServerNotificationService: CentralServerNotificationService,
     private centralServerService: CentralServerService,
+    private authorizationService: AuthorizationService,
     private datePipe: AppDatePipe) {
     super(spinnerService, translateService);
     this.isOrganizationComponentActive = this.componentService.isActive(TenantComponents.ORGANIZATION);
@@ -177,15 +179,22 @@ export class RegistrationTokensTableDataSource extends TableDataSource<Registrat
 
   public buildTableActionsDef(): TableActionDef[] {
     const tableActionsDef = super.buildTableActionsDef();
-    return [
-      new TableCreateAction().getActionDef(),
-      ...tableActionsDef,
-    ];
+    if (this.authorizationService.canCreateToken()) {
+      return [
+        new TableCreateAction().getActionDef(),
+        ...tableActionsDef,
+      ];
+    }
+    return tableActionsDef;
   }
 
   public buildTableDynamicRowActions(registrationToken: RegistrationToken): TableActionDef[] {
+    const rowActions = [];
     if (registrationToken.revocationDate || moment().isAfter(registrationToken.expirationDate)) {
-      return [this.deleteAction];
+      if (this.authorizationService.canDeleteToken()) {
+        return [this.deleteAction];
+      }
+      return [];
     }
     const copyUrlActions: TableActionDef[] = [
       ...(!Utils.isUndefined(registrationToken.ocpp15SOAPUrl) ? [this.copySOAP15Action] : []),
@@ -197,13 +206,15 @@ export class RegistrationTokensTableDataSource extends TableDataSource<Registrat
     ];
     this.copyUrlAction = new TableMultiCopyAction(
       copyUrlActions,
-      'settings.charging_station.copy_url_tooltip',
-      'settings.charging_station.copy_url_tooltip').getActionDef();
-    return [
-      this.copyUrlAction,
-      this.revokeAction,
-      this.deleteAction,
-    ];
+      'chargers.connections.copy_url_tooltip',
+      'chargers.connections.copy_url_tooltip').getActionDef();
+    
+    rowActions.push(this.copyUrlAction, this.revokeAction);
+
+    if (this.authorizationService.canDeleteToken()) {
+      rowActions.push(this.deleteAction)
+    }
+    return rowActions;
   }
 
   public actionTriggered(actionDef: TableActionDef) {
@@ -226,27 +237,27 @@ export class RegistrationTokensTableDataSource extends TableDataSource<Registrat
       case ButtonAction.COPY:
         let url;
         switch (actionDef.name) {
-          case 'settings.charging_station.ocpp_15_soap':
+          case 'chargers.connections.ocpp_15_soap':
             url = registrationToken.ocpp15SOAPUrl;
             break;
-          case 'settings.charging_station.ocpp_16_soap':
+          case 'chargers.connections.ocpp_16_soap':
             url = registrationToken.ocpp16SOAPUrl;
             break;
-          case 'settings.charging_station.ocpp_16_json':
+          case 'chargers.connections.ocpp_16_json':
             url = registrationToken.ocpp16JSONUrl;
             break;
-          case 'settings.charging_station.ocpp_15_soap_secure':
+          case 'chargers.connections.ocpp_15_soap_secure':
             url = registrationToken.ocpp15SOAPSecureUrl;
             break;
-          case 'settings.charging_station.ocpp_16_soap_secure':
+          case 'chargers.connections.ocpp_16_soap_secure':
             url = registrationToken.ocpp16SOAPSecureUrl;
             break;
-          case 'settings.charging_station.ocpp_16_json_secure':
+          case 'chargers.connections.ocpp_16_json_secure':
             url = registrationToken.ocpp16JSONSecureUrl;
             break;
         }
         Utils.copyToClipboard(url);
-        this.messageService.showInfoMessage('settings.charging_station.url_copied');
+        this.messageService.showInfoMessage('chargers.connections.url_copied');
         break;
     }
   }
@@ -278,21 +289,21 @@ export class RegistrationTokensTableDataSource extends TableDataSource<Registrat
 
   private deleteToken(registrationToken: RegistrationToken) {
     this.dialogService.createAndShowYesNoDialog(
-      this.translateService.instant('settings.charging_station.registration_token_delete_title'),
-      this.translateService.instant('settings.charging_station.registration_token_delete_confirm'),
+      this.translateService.instant('chargers.connections.registration_token_delete_title'),
+      this.translateService.instant('chargers.connections.registration_token_delete_confirm'),
     ).subscribe((result) => {
       if (result === ButtonType.YES) {
         this.centralServerService.deleteRegistrationToken(registrationToken.id).subscribe((response) => {
           if (response.status === RestResponse.SUCCESS) {
             this.refreshData().subscribe();
-            this.messageService.showSuccessMessage('settings.charging_station.registration_token_delete_success');
+            this.messageService.showSuccessMessage('chargers.connections.registration_token_delete_success');
           } else {
             Utils.handleError(JSON.stringify(response),
-              this.messageService, 'settings.charging_station.registration_token_delete_error');
+              this.messageService, 'chargers.connections.registration_token_delete_error');
           }
         }, (error) => {
           Utils.handleHttpError(error, this.router, this.messageService, this.centralServerService,
-            'settings.charging_station.registration_token_delete_error');
+            'chargers.connections.registration_token_delete_error');
         });
       }
     });
@@ -300,21 +311,21 @@ export class RegistrationTokensTableDataSource extends TableDataSource<Registrat
 
   private revokeToken(registrationToken: RegistrationToken) {
     this.dialogService.createAndShowYesNoDialog(
-      this.translateService.instant('settings.charging_station.registration_token_revoke_title'),
-      this.translateService.instant('settings.charging_station.registration_token_revoke_confirm'),
+      this.translateService.instant('chargers.connections.registration_token_revoke_title'),
+      this.translateService.instant('chargers.connections.registration_token_revoke_confirm'),
     ).subscribe((result) => {
       if (result === ButtonType.YES) {
         this.centralServerService.revokeRegistrationToken(registrationToken.id).subscribe((response) => {
           if (response.status === RestResponse.SUCCESS) {
             this.refreshData().subscribe();
-            this.messageService.showSuccessMessage('settings.charging_station.registration_token_revoke_success');
+            this.messageService.showSuccessMessage('chargers.connections.registration_token_revoke_success');
           } else {
             Utils.handleError(JSON.stringify(response),
-              this.messageService, 'settings.charging_station.registration_token_revoke_error');
+              this.messageService, 'chargers.connections.registration_token_revoke_error');
           }
         }, (error) => {
           Utils.handleHttpError(error, this.router, this.messageService, this.centralServerService,
-            'settings.charging_station.registration_token_revoke_error');
+            'chargers.connections.registration_token_revoke_error');
         });
       }
     });
