+++ src/app/pages/charging-stations/registration-tokens/registration-tokens.component.html
@@ -1,0 +1,12 @@
+<div class="card card-big">
+    <div class="card-header card-header-icon card-header-primary">
+      <div class="card-icon">
+        <mat-icon>link</mat-icon>
+      </div>
+      <h3 class="card-title">{{"chargers.connections.registration_tokens_title" | translate}}</h3>
+    </div>
+    <div class="card-body">
+      <app-table [dataSource]="registrationTokenDataSource"></app-table>
+    </div>
+  </div>
+
