+++ src/app/pages/charging-stations/registration-tokens/registration-tokens.component.ts
@@ -1,15 +1,16 @@
-import { Component } from '@angular/core';
+import { Component } from "@angular/core";
 
-import { RegistrationTokensTableDataSource } from './registration-tokens-table-data-source';
+import { RegistrationTokensTableDataSource } from "./registration-tokens-table-data-source";
 
 @Component({
-  selector: 'app-settings-ocpp',
+  selector: 'app-charging-stations-connection',
   templateUrl: 'registration-tokens.component.html',
   providers: [RegistrationTokensTableDataSource],
 })
 export class RegistrationTokensComponent {
   constructor(
     public registrationTokenDataSource: RegistrationTokensTableDataSource,
-  ) {
-  }
+    ) {
+    }
 }
+
