--- src/app/pages/settings/registration-tokens/registration-tokens.component.html
@@ -1,11 +1,0 @@
-<div class="card card-big">
-  <div class="card-header card-header-icon card-header-primary">
-    <div class="card-icon">
-      <mat-icon>link</mat-icon>
-    </div>
-    <h3 class="card-title">{{"settings.charging_station.registration_tokens_title" | translate}}</h3>
-  </div>
-  <div class="card-body">
-    <app-table [dataSource]="registrationTokenDataSource"></app-table>
-  </div>
-</div>
