+++ src/app/pages/settings/settings.component.ts
@@ -27,7 +27,7 @@ export class SettingsComponent extends AbstractTabComponent {
     activatedRoute: ActivatedRoute,
     windowService: WindowService,
   ) {
-    super(activatedRoute, windowService, ['ocpi', 'organization', 'refund', 'pricing', 'billing', 'sac', 'smartCharging', 'asset']);
+    super(activatedRoute, windowService, ['organization', 'refund', 'pricing', 'billing', 'sac', 'smartCharging', 'asset']);
     this.isOCPIActive = this.componentService.isActive(TenantComponents.OCPI);
     this.isOrganizationActive = this.componentService.isActive(TenantComponents.ORGANIZATION);
     this.isRefundActive = this.componentService.isActive(TenantComponents.REFUND);
