+++ src/assets/i18n/es.json
@@ -1004,7 +1004,8 @@
     "tabs": {
       "list": "Estaciones de carga",
       "charging_plans": "Planes de carga",
-      "in_error": "Con errores"
+      "in_error": "Con errores",
+      "connection": "Onboard New Station"
     },
     "errors": {
       "missing_settings": {
@@ -1027,6 +1028,37 @@
         "description": "La estación de carga envió un mensaje de error de uno de sus conectores",
         "action": "Verifique los detalles del error y reinicie la estación de carga"
       }
+    },
+    "connections": {
+      "registration_tokens_title": "Tokens de registro de la estación de carga",
+      "registration_token_site_area_name": "Token para la zona '{{siteAreaName}}'",
+      "registration_token_creation_title": "Crear token de registro",
+      "registration_token_creation_confirm": "¿Está seguro que desea crear un token de registro para las estaciones de carga de esta zona?",
+      "registration_token_creation_success": "El token de registro se ha creado exitosamente",
+      "registration_token_creation_error": "Se produjo un error al crear el token de registro",
+      "registration_token_revoke_title": "Revocar token de registro",
+      "registration_token_revoke_confirm": "¿Está seguro que desea revocar el token de registro?",
+      "registration_token_revoke_success": "El token de registro ha sido revocado exitosamente",
+      "registration_token_revoke_error": "Se produjo un error al revocar el token de registro",
+      "registration_token_delete_title": "Eliminar token de registro",
+      "registration_token_delete_confirm": "¿Está seguro que desea eliminar el token de registro?",
+      "registration_token_delete_success": "El token de registro se ha eliminado exitosamente",
+      "registration_token_delete_error": "Se produjo un error al eliminar el token de registro",
+      "registration_token_url": "URL de registro",
+      "registration_token_valid": "Válido",
+      "registration_token_expired": "Caducado",
+      "registration_token_revoked": "Revocado",
+      "token_restrict_site_area": "Restringir el acceso a una zona",
+      "url": "URL de OCPP",
+      "copy_url_tooltip": "Copiar URL de OCPP",
+      "url_copied": "La URL de OCPP se ha copiado en su portapapeles",
+      "generate_connection_url": "Conectar estación de carga",
+      "ocpp_15_soap": "OCPP 1.5 SOAP",
+      "ocpp_16_json": "OCPP 1.6 JSON",
+      "ocpp_16_soap": "OCPP 1.6 SOAP",
+      "ocpp_15_soap_secure": "Secure OCPP 1.5 SOAP",
+      "ocpp_16_json_secure": "Secure OCPP 1.6 JSON",
+      "ocpp_16_soap_secure": "Secure OCPP 1.6 SOAP"
     }
   },
   "logs": {
@@ -1315,7 +1347,6 @@
     "settings_not_saved": "Debe guardar la configuración primero",
     "tabs": {
       "ocpi": "Roaming",
-      "charging_station": "Estación de carga",
       "refund": "Reembolsos",
       "pricing": "Costos",
       "billing": "Facturas",
@@ -1334,37 +1365,6 @@
       "update_success": "La configuración del OCPI se ha actualizado correctamente",
       "update_error": "Se produjo un error al actualizar la configuración de OCPI",
       "description": "E-Mobility proporciona un servicio certificado para conectarse a una plataforma externa de E-Roaming. <br> Esta función expone en tiempo real la infraestructura y disponibilidad de las estaciones de carga a través del protocolo estándar OCPI (Interfaz de punto de carga abierta). <br> E-Mobility está utilizando actualmente la plataforma de interoperabilidad de Gireve (IOP)."
-    },
-    "charging_station": {
-      "registration_tokens_title": "Tokens de registro de la estación de carga",
-      "registration_token_site_area_name": "Token para la zona '{{siteAreaName}}'",
-      "registration_token_creation_title": "Crear token de registro",
-      "registration_token_creation_confirm": "¿Está seguro que desea crear un token de registro para las estaciones de carga de esta zona?",
-      "registration_token_creation_success": "El token de registro se ha creado exitosamente",
-      "registration_token_creation_error": "Se produjo un error al crear el token de registro",
-      "registration_token_revoke_title": "Revocar token de registro",
-      "registration_token_revoke_confirm": "¿Está seguro que desea revocar el token de registro?",
-      "registration_token_revoke_success": "El token de registro ha sido revocado exitosamente",
-      "registration_token_revoke_error": "Se produjo un error al revocar el token de registro",
-      "registration_token_delete_title": "Eliminar token de registro",
-      "registration_token_delete_confirm": "¿Está seguro que desea eliminar el token de registro?",
-      "registration_token_delete_success": "El token de registro se ha eliminado exitosamente",
-      "registration_token_delete_error": "Se produjo un error al eliminar el token de registro",
-      "registration_token_url": "URL de registro",
-      "registration_token_valid": "Válido",
-      "registration_token_expired": "Caducado",
-      "registration_token_revoked": "Revocado",
-      "token_restrict_site_area": "Restringir el acceso a una zona",
-      "url": "URL de OCPP",
-      "copy_url_tooltip": "Copiar URL de OCPP",
-      "url_copied": "La URL de OCPP se ha copiado en su portapapeles",
-      "generate_connection_url": "Conectar estación de carga",
-      "ocpp_15_soap": "OCPP 1.5 SOAP",
-      "ocpp_16_json": "OCPP 1.6 JSON",
-      "ocpp_16_soap": "OCPP 1.6 SOAP",
-      "ocpp_15_soap_secure": "Secure OCPP 1.5 SOAP",
-      "ocpp_16_json_secure": "Secure OCPP 1.6 JSON",
-      "ocpp_16_soap_secure": "Secure OCPP 1.6 SOAP"
     },
     "refund": {
       "description": "E-Mobility proporciona un servicio de reembolso para un sistema de gastos externo. <br> Esta característica permite al usuario llevar sus sesiones a un sistema de gastos y obtener un reembolso de la electricidad consumida. <br> E-Mobility usa actualmente SAP Concur. Más información técnica disponible en: ",
