+++ src/assets/i18n/fr.json
@@ -1004,7 +1004,8 @@
     "tabs": {
       "list": "Bornes",
       "charging_plans": "Charging Plans",
-      "in_error": "En Erreur"
+      "in_error": "En Erreur",
+      "connection": "Connecter une nouvelle borne"
     },
     "errors": {
       "missing_settings": {
@@ -1027,6 +1028,37 @@
         "description": "La borne a renvoyé un état en erreur sur un des connecteurs",
         "action": "Vérifier le code d'erreur et éventuellement redémarrer la borne"
       }
+    },
+    "connections": {
+      "registration_tokens_title": "Jetons d'enregistrement des bornes de recharge",
+      "registration_token_site_area_name": "Jetons d'enregistrement pour la zone '{{siteAreaName}}'",
+      "registration_token_creation_title": "Créer un jeton d'enregistrement",
+      "registration_token_creation_confirm": "Voulez-vous vraiment créer un jeton d'enregistrement pour les bornes de recharge de cette zone?",
+      "registration_token_creation_success": "Le jeton d'enregistrement a été créé avec succès",
+      "registration_token_creation_error": "Une erreur est survenue pendant la création du jeton d'enregistrement, vérifiez les logs",
+      "registration_token_revoke_title": "Révoquer un jeton d'enregistrement",
+      "registration_token_revoke_confirm": "Voulez-vous vraiment révoquer le jeton d'enregistrement ?",
+      "registration_token_revoke_success": "Le jeton d'enregistrement a été révoqué avec succès",
+      "registration_token_revoke_error": "Une erreur est survenue pendant la création du jeton d'enregistrement, vérifiez les logs",
+      "registration_token_delete_title": "Supprimer un jeton d'enregistrement",
+      "registration_token_delete_confirm": "Voulez-vous vraiment supprimer le jeton d'enregistrement ?",
+      "registration_token_delete_success": "Le jeton d'enregistrement a été supprimé avec succès",
+      "registration_token_delete_error": "Une erreur est survenue pendant la suppression du jeton d'enregistrement, vérifiez les logs",
+      "registration_token_url": "URL d'enregistrement",
+      "registration_token_valid": "Valide",
+      "registration_token_expired": "Expiré",
+      "registration_token_revoked": "Révoqué",
+      "token_restrict_site_area": "Restreindre l'accès à une zone",
+      "url": "URL OCPP",
+      "copy_url_tooltip": "Copier l'URL OCPP",
+      "url_copied": "L'URL OCPP a été copiée dans votre presse-papiers",
+      "generate_connection_url": "Connecter une Borne",
+      "ocpp_15_soap": "OCPP 1.5 SOAP",
+      "ocpp_16_json": "OCPP 1.6 JSON",
+      "ocpp_16_soap": "OCPP 1.6 SOAP",
+      "ocpp_15_soap_secure": "OCPP 1.5 SOAP Sécurisé",
+      "ocpp_16_json_secure": "OCPP 1.6 JSON Sécurisé",
+      "ocpp_16_soap_secure": "OCPP 1.6 SOAP Sécurisé"
     }
   },
   "logs": {
@@ -1315,7 +1347,6 @@
     "settings_not_saved": "Vous devez d'abord enregistrer vos changements",
     "tabs": {
       "ocpi": "Itinérance",
-      "charging_station": "Borne",
       "refund": "Remboursement",
       "pricing": "Tarification",
       "billing": "Facturation",
@@ -1334,37 +1365,6 @@
       "update_success": "La configuration OCPI a été mis à jour avec succès",
       "update_error": "Une erreur est survenue pendant la mise à jour de la configuration OCPI, vérifiez les logs",
       "description": "E-Mobility fournit un service certifié permettant de se connecter à une plate-forme d'itinérance externe.<br>Cette fonctionnalité permet de visualiser en temps réel l’infrastructure et la disponibilité des bornes de recharge via le protocole standard OCPI (Open Charge Point Interface).<br>E-Mobility utilise actuellement la plate-forme d'interopérabilité Gireve (IOP)."
-    },
-    "charging_station": {
-      "registration_tokens_title": "Jetons d'enregistrement des bornes de recharge",
-      "registration_token_site_area_name": "Token pour la zone '{{siteAreaName}}'",
-      "registration_token_creation_title": "Créer un jeton d'enregistrement",
-      "registration_token_creation_confirm": "Voulez-vous vraiment créer un jeton d'enregistrement pour les bornes de recharge de cette zone?",
-      "registration_token_creation_success": "Le jeton d'enregistrement a été créé avec succès",
-      "registration_token_creation_error": "Une erreur est survenue pendant la création du jeton d'enregistrement, vérifiez les logs",
-      "registration_token_revoke_title": "Révoquer un jeton d'enregistrement",
-      "registration_token_revoke_confirm": "Voulez-vous vraiment révoquer le jeton d'enregistrement ?",
-      "registration_token_revoke_success": "Le jeton d'enregistrement a été révoqué avec succès",
-      "registration_token_revoke_error": "Une erreur est survenue pendant la création du jeton d'enregistrement, vérifiez les logs",
-      "registration_token_delete_title": "Supprimer un jeton d'enregistrement",
-      "registration_token_delete_confirm": "Voulez-vous vraiment supprimer le jeton d'enregistrement ?",
-      "registration_token_delete_success": "Le jeton d'enregistrement a été supprimé avec succès",
-      "registration_token_delete_error": "Une erreur est survenue pendant la suppression du jeton d'enregistrement, vérifiez les logs",
-      "registration_token_url": "URL d'enregistrement",
-      "registration_token_valid": "Valide",
-      "registration_token_expired": "Expiré",
-      "registration_token_revoked": "Révoqué",
-      "token_restrict_site_area": "Restreindre l'accès à une zone",
-      "url": "OCPP URL",
-      "copy_url_tooltip": "Copier l'URL OCPP",
-      "url_copied": "L'URL OCPP a été copiée dans votre presse-papiers",
-      "generate_connection_url": "Connecter une Borne",
-      "ocpp_15_soap": "OCPP 1.5 SOAP",
-      "ocpp_16_json": "OCPP 1.6 JSON",
-      "ocpp_16_soap": "OCPP 1.6 SOAP",
-      "ocpp_15_soap_secure": "OCPP 1.5 SOAP Sécurisé",
-      "ocpp_16_json_secure": "OCPP 1.6 JSON Sécurisé",
-      "ocpp_16_soap_secure": "OCPP 1.6 SOAP Sécurisé"
     },
     "refund": {
       "description": "E-Mobility fournit un service de remboursement pour des systèmes externes.<br>Cette fonctionnalité permet à l'utilisateur de se faire rembourser des dépenses d'électricité liées à ses transactions.<br>E-Mobility utilise actuellement SAP Concur.<br>Plus d'informations techniques disponibilite sur ",
